+++ src/main/java/ru/cft/shift/intensive/template/service/PurchaseService.java
@@ -22,5 +22,5 @@ public interface PurchaseService {
 
     Float CalculateTotalCost();
 
-    Float CalculateDeliveryTime();
+    Integer CalculateDeliveryTime();
 }
